fix: checked for user before updating and deleting

+++ controllers/user.controller.js
@@ -2,7 +2,6 @@ import pool from '../dbConfig.mjs';
 
 export const createUser = async (req, res, next) => {
   try {
-    console.log(req.body);
     const { username, age, email } = req.body;
     const query = `
       INSERT INTO users (username, age, email)
@@ -24,17 +23,30 @@ export const updateUser = async (req, res, next) => {
   try {
     const { id } = req.params;
     const { username, age, email } = req.body;
-    const query = `
+
+    const userExistsQuery = `
+      SELECT * FROM users
+      WHERE id = $1;
+    `;
+    const userExistsValues = [id];
+    const userExistsResult = await pool.query(userExistsQuery, userExistsValues);
+
+    if (userExistsResult.rows.length === 0) {
+      return res.status(404).json({ message: 'User not found.' });
+    }
+
+    const updateUserQuery = `
       UPDATE users
       SET username = $1, age = $2, email = $3
       WHERE id = $4
       RETURNING *;
     `;
-    const values = [username, age, email, id];
-    const result = await pool.query(query, values);
+    const updateUserValues = [username, age, email, id];
+    const result = await pool.query(updateUserQuery, updateUserValues);
+
     res.json({
-      "message": "User updated successfully.",
-      "user": result.rows[0]
+      message: 'User updated successfully.',
+      user: result.rows[0]
     });
   } catch (error) {
     next(error);
@@ -44,15 +56,28 @@ export const updateUser = async (req, res, next) => {
 export const deleteUser = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const query = `
+
+    const userExistsQuery = `
+      SELECT * FROM users
+      WHERE id = $1;
+    `;
+    const userExistsValues = [id];
+    const userExistsResult = await pool.query(userExistsQuery, userExistsValues);
+
+    if (userExistsResult.rows.length === 0) {
+      return res.status(404).json({ message: 'User not found.' });
+    }
+
+    const deleteUserQuery = `
       DELETE FROM users
       WHERE id = $1;
     `;
-    const values = [id];
-    await pool.query(query, values);
+    const deleteUserValues = [id];
+    await pool.query(deleteUserQuery, deleteUserValues);
+
     res.json({ message: 'User deleted successfully.' });
   } catch (error) {
-   next(error);
+    next(error);
   }
 };
 
@@ -63,10 +88,12 @@ export const getUsers = async (req, res, next) => {
     `;
     const result = await pool.query(query);
     res.json({
-      "message": "Users retrieved  successfully.",
-      "user": result.rows[0]
+      "message": "Users retrieved successfully.",
+      "users": result.rows 
     });
   } catch (error) {
     next(error);
   }
 };
+
+
